+++ test/cypress/support/R4/assert-sdc-questionnaire.js
@@ -5,9 +5,9 @@ function assertFHTQuestionnaire(fhirData, fhirVersion='R4') {
   expect(fhirData.title).toBe("USSG-FHT, (with mock-up items for skip logic demo)");
   if (fhirVersion === 'R4B')
     expect(fhirData.meta.profile).toEqual(['http://hl7.org/fhir/4.3/StructureDefinition/Questionnaire']);
-  else 
+  else
     expect(fhirData.meta.profile).toEqual(['http://hl7.org/fhir/4.0/StructureDefinition/Questionnaire']);
-  
+
   // No URL conversion for now. -Ajay
   //expect(fhirData.url).toBe("http://hl7.org/fhir/us/sdc/Questionnaire/54127-6N");
   expect(fhirData.code[0].code).toBe("54127-6N");
@@ -21,7 +21,6 @@ function assertFHTQuestionnaire(fhirData, fhirVersion='R4') {
   expect(fhirData.item[0].code[0].display).toBe("Your health information");
   expect(fhirData.item[0].code[0].system).toBe(undefined);
   expect(fhirData.item[0].text).toBe("Your health information");
-  expect(fhirData.item[0].required).toBe(false);
   expect(fhirData.item[0].linkId).toBe("/54126-8");
   expect(fhirData.item[0].type).toBe("group");
 
@@ -29,7 +28,6 @@ function assertFHTQuestionnaire(fhirData, fhirVersion='R4') {
 
   expect(fhirData.item[0].item[0].text).toBe("Name");
   expect(fhirData.item[0].item[0].type).toBe("text");
-  expect(fhirData.item[0].item[0].required).toBe(false);
   expect(fhirData.item[0].item[0].repeats).toBe(true);
   expect(fhirData.item[0].item[0].linkId).toBe("/54126-8/54125-0");
   expect(fhirData.item[0].item[0].code[0].code).toBe("54125-0");
@@ -38,7 +36,6 @@ function assertFHTQuestionnaire(fhirData, fhirVersion='R4') {
 
   expect(fhirData.item[0].item[12].text).toBe("Your diseases history");
   expect(fhirData.item[0].item[12].type).toBe("group");
-  expect(fhirData.item[0].item[12].required).toBe(false);
   expect(fhirData.item[0].item[12].repeats).toBe(true);
   expect(fhirData.item[0].item[12].linkId).toBe("/54126-8/54137-5");
   expect(fhirData.item[0].item[12].code[0].code).toBe("54137-5");
@@ -48,7 +45,6 @@ function assertFHTQuestionnaire(fhirData, fhirVersion='R4') {
 
   expect(fhirData.item[0].item[12].item[0].text).toBe("Disease or Condition");
   expect(fhirData.item[0].item[12].item[0].type).toBe("choice");
-  expect(fhirData.item[0].item[12].item[0].required).toBe(false);
   expect(fhirData.item[0].item[12].item[0].repeats).toBeFalsy();
   expect(fhirData.item[0].item[12].item[0].linkId).toBe("/54126-8/54137-5/54140-9");
   expect(fhirData.item[0].item[12].item[0].code[0].code).toBe("54140-9");
@@ -65,7 +61,6 @@ function assertFHTQuestionnaire(fhirData, fhirVersion='R4') {
   expect(fhirData.item[1].code[0].display).toBe("Family member health information");
   expect(fhirData.item[1].code[0].system).toBe(undefined);
   expect(fhirData.item[1].text).toBe("Family member health information");
-  expect(fhirData.item[1].required).toBe(false);
   expect(fhirData.item[1].linkId).toBe("/54114-4");
   expect(fhirData.item[1].type).toBe("group");
 }